Cleaning up resources. Adding images for Bands and Bracelets

+++ .scripts/Helpers/index.ts
@@ -1,6 +1,9 @@
 import Handlebars from 'handlebars';
-import { mkdirSync, readFileSync, writeFileSync } from 'fs';
-import { resolve, dirname } from 'path';
+import {mkdirSync, readFileSync, writeFileSync} from 'fs';
+import {dirname, resolve} from 'path';
+import {options} from "../index";
+
+export const projectDir = resolve(__dirname, '..', '..') + '\\';
 
 export function compileTemplate(baseDir: string, template: string) {
     return Handlebars.compile(readFileSync(resolve(baseDir, template), 'utf8'));
@@ -8,6 +11,9 @@ export function compileTemplate(baseDir: string, template: string) {
 
 export function writeGeneratedFile(dir: string, fileName: string, content: string) {
     const fullPath = resolve(dir, fileName);
-    mkdirSync(dirname(fullPath), { recursive: true });
+    if (options.verbose) console.log((options.dryRun ? '[DRY-RUN] ' : '') + 'Generated file: ' + fullPath.replace(projectDir, ''));
+    if (options.dryRun) return;
+
+    mkdirSync(dirname(fullPath), {recursive: true});
     writeFileSync(fullPath, content, 'utf8');
 }
